requirements.txt
numpy==1.19.2
XlsxWriter==1.3.7
pyzbar==0.1.8
opencv_python==4.4.0.46
protobuf==3.14.0
google-cloud-storage==1.35.0
google-cloud-vision==2.0.0